ajustando organizacao e alterando congiguracao do card
