+++ src/modules/home/home.styled.tsx
@@ -1,31 +1,11 @@
 import styled from "styled-components";
-import { Color, Padding } from "../components/obj.connstant/constant";
-
-export const Background = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  background-image: linear-gradient(
-    to right,
-    #272929,
-    #50298ce3,
-    #15053c,
-    #330654
-  );
-`;
+import { Color, Padding } from "../../components/obj.connstant";
 
 export const CardsSection = styled.section`
   display: flex;
   flex-wrap: wrap;
   gap: 20px;
   padding: ${Padding.Small};
-`;
-
-export const WrapperMoreInfo = styled.section`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
 `;
 
 export const WrapperSearch = styled.div`
