+++ src/modules/information/more-info.style.tsx
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+
+export const WrapperMoreInfo = styled.section`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+`;
