+++ src/modules/style.tsx
@@ -1,0 +1,14 @@
+import styled from "styled-components";
+
+export const Background = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  background-image: linear-gradient(
+    to right,
+    #272929,
+    #50298ce3,
+    #15053c,
+    #330654
+  );
+`;
